\name{RoundOff-package}
\alias{RoundOff-package}
\alias{RoundOff}
\docType{package}
\title{RoundOff
\packageTitle{RoundOff}
}
\description{Addresses problem with Base R's round() function, replacing it with a function which always rounds up for values ≥ 5. Base R's round() will always round to the nearest even, so 3.5 rounds to 4, but 4.5 also rounds to 4. 
\packageDescription{RoundOff}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RoundOff}
\packageIndices{RoundOff}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{Martin Ingram
\packageAuthor{RoundOff}

Maintainer: \packageMaintainer{RoundOff}
}
\references{https://stackoverflow.com/questions/12688717/round-up-from-5
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
